+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fixed issue when it was not possible to copy long text as it was folding/unfolding on this action, #320
 * Fixed issue when newlines and some of the other issue when user was posing comment as non-admin user
 * Fixed issue when global border radius was not changing based on the settings value
+* Fixed issue when repeating error messages could have been shown in case user tried to use admin's email to login via social, #312
 
 ## 0.0.72 – 15.11.2018