+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fixed issue when newlines and some of the other issue when user was posing comment as non-admin user
 * Fixed issue when global border radius was not changing based on the settings value
 * Fixed issue when repeating error messages could have been shown in case user tried to use admin's email to login via social, #312
+* Possible fix to the issue when some of the users experienced problems with real time notification about new comments, #310
 
 **Other:**